pyproject.toml
[project]
name = "musicreco"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.4",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "python-amazon-paapi>=5.0.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.38.0",
]
